fix(articleparamsform, app, index): перенесен ArticleParamsForm, вынесен отдельно App, мелкие фиксы
Состояние формы описано в самом компоненте ArticleParamsForm, для настроек по умолчанию использован
defaultArticleState, поправлена типизация, вынесен компонент App в отдельный файл, убрано
использование clsx для одиночных селекторов.

Review #1
